+++ paper/paper.bib
@@ -1,0 +1,77 @@
+@incollection{FRENKEL200223,
+title = "Chapter 3 - Monte Carlo Simulations",
+editor = "Daan Frenkel and Berend Smit",
+booktitle = "Understanding Molecular Simulation (Second Edition)",
+publisher = "Academic Press",
+edition = "Second Edition",
+address = "San Diego",
+pages = "23 - 61",
+year = "2002",
+isbn = "978-0-12-267351-1",
+doi = "https://doi.org/10.1016/B978-012267351-1/50005-5",
+url = "http://www.sciencedirect.com/science/article/pii/B9780122673511500055",
+author = "Daan Frenkel and Berend Smit"
+}
+@incollection{FRENKEL200263,
+title = "Chapter 4 - Molecular Dynamics Simulations",
+editor = "Daan Frenkel and Berend Smit",
+booktitle = "Understanding Molecular Simulation (Second Edition)",
+publisher = "Academic Press",
+edition = "Second Edition",
+address = "San Diego",
+pages = "63 - 107",
+year = "2002",
+isbn = "978-0-12-267351-1",
+doi = "https://doi.org/10.1016/B978-012267351-1/50006-7",
+url = "http://www.sciencedirect.com/science/article/pii/B9780122673511500067",
+author = "Daan Frenkel and Berend Smit"
+}
+
+@conference{KLUYVER,
+	Author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing},
+	Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
+	Editor = {F. Loizides and B. Schmidt},
+	Organization = {IOS Press},
+	Pages = {87 - 90},
+	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
+	Year = {2016}
+}
+
+@article{HUNTER,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing In Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python
+  for application development, interactive scripting, and
+  publication-quality image generation across user
+  interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
+
+@article {LENNARD-JONES,
+	title = {On the determination of molecular fields. {\textemdash}II. From the equation of state of a gas},
+	volume = {106},
+	number = {738},
+	pages = {463--477},
+	year = {1924},
+	doi = {10.1098/rspa.1924.0082},
+	publisher = {The Royal Society},
+	issn = {0950-1207},
+	URL = {http://rspa.royalsocietypublishing.org/content/106/738/463},
+	eprint = {http://rspa.royalsocietypublishing.org/content/106/738/463.full.pdf},
+	journal = {Proceedings of the Royal Society of London A: Mathematical, Physical and Engineering Sciences}
+}
+
+@article{FIGURE1,
+author = "Andrew McCluskey",
+title = "{fig1}",
+year = "2018",
+month = "6",
+url = "https://figshare.com/articles/fig1/6571583",
+doi = "10.6084/m9.figshare.6571583.v1"
+}
