### 2018-06-25 #17 @labarba
Hello, @arm61 -- Your submission is lacking the required paper. Please see the [Author Guidelines](http://jose.theoj.org/about#author_guidelines).
