### 2018-06-25 #17 @arm61
Hey @labarba, just realised that the paper directory was still in my .gitignore, resolved in https://github.com/arm61/pylj/commit/ef0e4dcfbf78aea02c17c6a0e3cb7b5198f5ebcd

sorry about that.
