### 2018-06-25 #17 @arm61
Title of paper wasn't in inverted commas, fingers crossed for https://github.com/arm61/pylj/commit/3ce1332e0a829e6656352baa6ea4228783826279
