### 2018-06-25 #17 @arm61
Possible reviewers from those already agreed to review of JOSE would include @shivupa, @william-pfalzgraff, and @TJFord. Although I am sure there are many other qualified individuals.
