### 2018-06-25 #17 @shivupa
Hi! I'd like to confirm that I am eligible to be a reviewer as I am only a graduate student. I do have publications on [related topics.](https://orcid.org/0000-0002-8501-0501)
