### 2018-06-26 #17 @labarba
@arm61 -- Thanks for fixing that.

Now, I understand that this software implements basic molecular dynamics algorithms. And I see that you have two Jupyter notebooks in the `examples` folder (molecular dynamics, and Monte Carlo). If you read the [Author Guid](http://jose.theoj.org/about#author_guidelines), you'll see that we are expecting "learning modules"—that is, complete worked examples with narration.

Your software may be useful by itself as a teaching tool, but it will be much more useful with full worked-out examples.
