### 2018-06-26 #17 @arm61
I had actually hoped to submit pylj as a piece of educational open-source software in its own right. And then once the laboratory exercises that we are developing within the University of Bath (which make use of pylj) are fully baked these would be submitted as full learning modules (indeed examples/ideal_gas_law/ideal_gas_law.ipynb is an example of one of such exercise).

I would be happy to flesh out the MC and MD examples if necessary to enable submission as an educational open-source software
