### 2018-06-26 #17 @labarba
I realize that. The first submissions to JOSE have turned out to be different than we were anticipating in the founding of the journal: our Call for Papers talks about two kinds of submissions: "learning modules"—which are fully narrated worked examples that use code to teach a technical subject—and "educational software"—which are infrastructure-type software, like auto-graders.

We've started to get submission that are just software, but not infrastructure software. That is, the "code part" of a learning module that uses coding to teach. It can be acceptable in some cases, but for the most part, narrated worked out examples are necessary for another instructor to be realistically able to adopt the software in their teaching, or for a self-learner to use the software to teach themselves.
