### 2018-06-26 #17 @arm61
Ahh right, I see now. In hindsight I recognise that wording in the Authors Guide, however, it is not initially crystal clear. :smile:

Before the weekend I will endeavour to flesh out the examples such that someone would be able to easily follow them to understand what the software is capable of and let you know when the appropriate commit is there. Would this be acceptable?
