+++ TODO.md
@@ -1,8 +1,9 @@
 ## todo
 
-- improve citation methods
-- testing for md sample function
-- testing for sample module
-- add energy minimisation
-- more example lesssons
-- find better solution/funding for online resource
+-[ ] improve the examples such that the explain the utility and use of pylj
+-[ ] improve citation methods
+-[ ] testing for md sample function
+-[ ] testing for sample module
+-[ ] add energy minimisation
+-[ ] more example lesssons
+-[ ] find better solution/funding for online resource
