### 2018-06-26 #17 @labarba
@TJFord You can find the paper PDF via the link provided by our helpful bot, whedon: https://github.com/openjournals/jose-reviews/issues/17#issuecomment-400088464

If you agree to review, yes, you are asked to review the software elements in the submission repository. Have a look at the [Reviewer Guidelines](http://jose.theoj.org/about#reviewer_guidelines). Once the review starts (this thread is the "pre-review"), you will have a handy checklist. Maybe look at this recent completed review: https://github.com/openjournals/jose-reviews/issues/15
