### 2018-06-26 #17 @TJFord
Hello Lorena,

I would like to review the paper. Sorry for the late response. I don't know
exactly how this works.  I only reviewed for traditional journals. I will
look at your posted links and work on that.

Jifu

On Tue, Jun 26, 2018 at 12:19 PM, Lorena A. Barba <notifications@github.com>
wrote:

> Thanks, @william-pfalzgraff <https://github.com/william-pfalzgraff>
> —we've not had confirmation from @TJFord <https://github.com/TJFord> that
> he accepts the reviewer role. Stay tuned.
>
