### 2018-06-26 #19 @whedon
[REVIEW]: pylj: A teaching tool for classical atomistic simulation
**Submitting author:** @arm61 (<a href="http://orcid.org/0000-0003-3381-5911">Andrew McCluskey</a>)
**Repository:** <a href="https://github.com/arm61/pylj" target ="_blank">https://github.com/arm61/pylj</a>
**Version:** v1.0.0
