### 2018-06-26 #17 @labarba
@william-pfalzgraff Thank you for your gracious volunteering! I'll be sure reaching out to you soon with another submission. In any case, if you're curious, the Review issue is open for comments from anyone. (Follow the link above to the Review issue.)
