+++ paper/paper.bib
@@ -8,7 +8,7 @@ address = "San Diego",
 pages = "23 - 61",
 year = "2002",
 isbn = "978-0-12-267351-1",
-doi = "https://doi.org/10.1016/B978-012267351-1/50005-5",
+doi = "10.1016/B978-012267351-1/50005-5",
 url = "http://www.sciencedirect.com/science/article/pii/B9780122673511500055",
 author = "Daan Frenkel and Berend Smit"
 }
@@ -22,7 +22,7 @@ address = "San Diego",
 pages = "63 - 107",
 year = "2002",
 isbn = "978-0-12-267351-1",
-doi = "https://doi.org/10.1016/B978-012267351-1/50006-7",
+doi = "10.1016/B978-012267351-1/50006-7",
 url = "http://www.sciencedirect.com/science/article/pii/B9780122673511500067",
 author = "Daan Frenkel and Berend Smit"
 }
