### 2018-06-26 a740594
Added doi to jupyter notebooks reference
