+++ paper/paper.bib
@@ -33,6 +33,7 @@ author = "Daan Frenkel and Berend Smit"
 	Editor = {F. Loizides and B. Schmidt},
 	Organization = {IOS Press},
 	Pages = {87 - 90},
+	doi = {10.3233/978-1-61499-649-1-87},
 	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
 	Year = {2016}
 }
