+++ README.md
@@ -19,7 +19,7 @@ python setup.py test
 All development discussion can take place on [gitter](https://gitter.im/pylj/Lobby#).
 
 [![DOI](https://zenodo.org/badge/119863480.svg)](https://zenodo.org/badge/latestdoi/119863480)
-[![Documentation Status](https://readthedocs.org/projects/falass/badge/?version=latest)](https://falass.readthedocs.io/en/latest/?badge=latest)
+[![Documentation Status](https://readthedocs.org/projects/pylj/badge/?version=latest)](http://pylj.readthedocs.io/en/latest/?badge=latest)
 [![PyPI version](https://badge.fury.io/py/pylj.svg)](https://badge.fury.io/py/pylj)
 [![Build Status](https://travis-ci.org/arm61/pylj.svg?branch=master)](https://travis-ci.org/arm61/pylj)
 [![Coverage Status](https://coveralls.io/repos/github/arm61/pylj/badge.svg?branch=master)](https://coveralls.io/github/arm61/pylj?branch=master)
