### 2018-06-26 #19 @labarba
Isn't our bot `whedon` just the cutest thing?
