### 2018-06-26 #19 @TJFord
A few comments:
1.  I really like the idea of running a demo through Jupyter, as it is convenient to test the code directly. I tried the cloud version for molecular_dynamics but not successful. Please correct me if I did it wrong. http://35.230.133.1/notebook/notebooks/examples/molecular_dynamics.ipynb#
The error message is here:

<pre>TypeError                                 Traceback (most recent call last)
<ipython-input-17-980173a92a59> in <module>()
----> 1 system = md_simulation(100, 273.15, 100, 5000, 50)
<ipython-input-16-0578d5628985> in md_simulation(number_of_particles, temperature, box_length, number_of_steps, sample_frequency)
     16                                                                                                  system.cut_off)
     17         # Run the equations of motion integrator algorithm
---> 18         system.particles = md.velocity_verlet(system.particles, system.timestep_length, system.box_length)
     19         # Sample the thermodynamic and structural parameters of the system
     20         system = md.sample(system.particles, system.box_length, system.initial_particles, system)
TypeError: velocity_verlet() missing 1 required positional argument: 'cut_off' </pre>

2. I tried to run it on a linux computer through a terminal, it was not successful either. I followed exactly all the steps shown in README.md.  I suspected I missed installing some libraries. I think it will be beneficial if a detailed installation manual can be provided.
