### 2018-06-26 b17af7d
update contributing and readme to reflect the move away from slack to gitter
