### 2018-06-26 #19 @TJFord
@shivupa I just opened an issue at https://github.com/arm61/pylj/issues/11
