### 2018-06-26 arm61/pylj#11 @arm61
1) has now been resolved and the online resource updated. so the molecular_dynamics notebook should run.
