### 2018-06-26 arm61/pylj#11 @arm61
Regarding 2. Are you trying to run components of the library? as all visualisation functions, e.g. those in the `sample` module require Jupyter notebooks to function. Really the whole software is designed to work within the Jupyter notebooks framework to allow for the easy visualisation to work. I have updated the README.md to reflect the requirement for Jupyter notebooks in https://github.com/arm61/pylj/commit/90de7f99acdd2d5122635286527faa2875f8a8b6
