### 2018-06-26 arm61/pylj#11 @shivupa
Also have you considered [binder](https://mybinder.org/) or [colaboratory](https://research.google.com/colaboratory/faq.html)
