### 2018-06-26 arm61/pylj#11 @arm61
I don't believe so (the delay is mentioned on the [webpages](http://pythoninchemistry.org/pylj/#examples)), the online resource is designed for quick demonstration purposes only. Currently this is running on free credit on a GCloud server and I don't really have the funds to pay for it to run more instances currently.
