### 2018-06-26 arm61/pylj#11 @arm61
I previously had it running on a GCloud binder resource (not mybinder) however as it is only really for quick demonstration purposes currently I found it was easier to manage in its current incarnation
