### 2018-06-30 #19 @TJFord
@labarba  @shivupa it looks like the example monte carlo simulation is not working either. @arm61, can you double check it?
Also a few other minor suggestions:
1. would it be better if the title is changed to "pylj: A teaching tool for classical molecular dynamics simulation "? as the focus of the tool is on molecular dynamics simulation.
2. I would like to cite a few other widely used open source molecular dynamics simulation packages in the paper, such as lammps, gromacs, etc., for those students if they want to know more about it.

<pre>---------------------------------------------------------------------------
<ipython-input-22-4859d5764793> in <module>()
----> 1 system = mc_simulation(2, 273.15, 45, 100000, 100)

<ipython-input-21-456fcaae123c> in mc_simulation(number_of_particles, temperature, box_length, number_of_steps, sample_frequency)
      8     system.particles, system.distances, system.energies = comp.compute_energy(system.particles,
      9                                                                               system.box_length,
---> 10                                                                               system.cut_off)
     11     old_energy = system.energies.sum()
     12     system = mc.sample(system.particles, system.box_length, old_energy, system)

AttributeError: 'System' object has no attribute 'cut_off'</pre>
