### 2018-07-01 #19 @shivupa
I'd suggest "molecular simulation" since this encompasses molecular dynamics and Monte Carlo.
