### 2018-07-01 arm61/pylj#11 @shivupa
Sorry for the delay. For the submission, the way it's set up is fine. A possible overhaul of the online part could be something to address down the line.
