### 2018-07-01 arm61/pylj#11 @shivupa
@TJFord  are you still having trouble? Both notebooks run for me now. I'd be ok with closing this.
