### 2018-07-02 arm61/pylj#11 @arm61
This has now been updated and should work (the server was not updated with the newest instance of the monte_carlo notebook)
