### 2018-07-02 #19 @arm61
Can we possibly agree on "classical atomistic simulation"? For the following reasons:

- Since we are only really dealing with individual atoms of argon in pylj it is a bit false to say molecular.
- I would like to keep "classical" in there as pylj only performs classical potential based simulation, e.g. there are not quantum mechanical methods (e.g. DTF) to be seen.
