### 2018-07-02 f5c4d95
getting rid of the magic numbers in the code and providing some rationale for them via comments.
