+++ pylj/mc.py
@@ -150,7 +150,8 @@ def metropolis(temperature, old_energy, new_energy, n = np.random.rand()):
     bool
         True if the move should be accepted.
     """
-    beta = 1 / (1.3806e-23 * temperature)
+    boltzmann_constant = 1.3806e-23 # joules/kelvin
+    beta = 1 / (boltzmann_constant * temperature)
     energy_difference = new_energy - old_energy
     metropolis_factor = np.exp(-beta * energy_difference)
     if n < metropolis_factor:
