### 2018-07-02 f8381bf
improving efficiency slightly
