### 2018-07-02 6131a7b
simplifaction of the examples by the mapping of many function within the system class
