### 2018-07-02 3441d7d
improving efficiency slightly
