+++ pylj/md.py
@@ -1,5 +1,4 @@
 import numpy as np
-from pylj import util
 try:
     from pylj import comp as heavy
 except ImportError:
@@ -31,6 +30,7 @@ def initialise(number_of_particles, temperature, box_length, init_conf, timestep
     System
         System information.
     """
+    from pylj import util
     system = util.System(number_of_particles, temperature, box_length, init_conf=init_conf,
                          timestep_length=timestep_length)
     v = np.random.rand(system.particles.size, 2) - 0.5
@@ -106,15 +106,43 @@ def sample(particles, box_length, initial_particles, system):
         Details about the whole system, with the new temperature, pressure, msd, and force appended to the appropriate
         arrays.
     """
-    temperature_new = util.calculate_temperature(particles)
+    temperature_new = calculate_temperature(particles)
     system.temperature_sample = np.append(system.temperature_sample, temperature_new)
     pressure_new = heavy.calculate_pressure(particles, box_length, temperature_new, system.cut_off)
-    msd_new = util.calculate_msd(particles, initial_particles, box_length)
+    msd_new = calculate_msd(particles, initial_particles, box_length)
     system.pressure_sample = np.append(system.pressure_sample, pressure_new)
     system.force_sample = np.append(system.force_sample, np.sum(system.forces))
     system.energy_sample = np.append(system.energy_sample, np.sum(system.energies))
     system.msd_sample = np.append(system.msd_sample, msd_new)
     return system
+
+
+def calculate_msd(particles, initial_particles, box_length):
+    """Determines the mean squared displacement of the particles in the system.
+
+    Parameters
+    ----------
+    particles: util.particle_dt, array_like
+        Information about the particles.
+    initial_particles: util.particle_dt, array_like
+        Information about the initial state of the particles.
+    box_length: float
+        Size of the cell vector.
+
+    Returns
+    -------
+    float:
+        Mean squared deviation for the particles at the given timestep.
+    """
+    dx = particles['xposition'] - initial_particles['xposition']
+    dy = particles['yposition'] - initial_particles['yposition']
+    for i in range(0, particles['xposition'].size):
+        if np.abs(dx[i]) > 0.5 * box_length:
+            dx[i] *= 1 - box_length / np.abs(dx[i])
+        if np.abs(dy[i]) > 0.5 * box_length:
+            dy[i] *= 1 - box_length / np.abs(dy[i])
+    dr = np.sqrt(dx * dx + dy * dy)
+    return np.average(dr ** 2)
 
 
 def update_positions(positions, velocities, accelerations, timestep_length, box_length):
@@ -167,4 +195,29 @@ def update_velocities(velocities, accelerations_old, accelerations_new, timestep
     velocities[0] += 0.5 * (accelerations_old[0] + accelerations_new[0]) * timestep_length
     velocities[1] += 0.5 * (accelerations_old[1] + accelerations_new[1]) * timestep_length
     return [velocities[0], velocities[1]]
+
+
+def calculate_temperature(particles):
+    """Determine the instantaneous temperature of the system.
+
+    Parameters
+    ----------
+    particles: util.particle_dt, array_like
+        Information about the particles.
+
+    Returns
+    -------
+    float:
+        Calculated instantaneous simulation temperature.
+    """
+    k = 0
+    for i in range(0, particles['xposition'].size):
+        v = np.sqrt(particles['xvelocity'][i] * particles['xvelocity'][i] + particles['yvelocity'][i] *
+                    particles['yvelocity'][i])
+        boltzmann_constant = 1.3806e-23  # joules/kelvin
+        atomic_mass_unit = 1.660539e-27 # kilograms
[…]
