### 2018-07-02 fc9dbf1
fix tests
