### 2018-07-03 90b8add
adding tests to pairwise
