### 2018-07-03 bb492be
Update paper.md
