### 2018-07-03 #19 @arm61
Paper title has been updated in https://github.com/arm61/pylj/commit/bb492bed8d8d37656e2534574f8ba4a876d001a6
