### 2018-07-03 f5b350d
add some tests for comp
