+++ pylj.egg-info/SOURCES.txt
@@ -15,6 +15,7 @@ pylj.egg-info/not-zip-safe
 pylj.egg-info/requires.txt
 pylj.egg-info/top_level.txt
 pylj/tests/__init__.py
+pylj/tests/test_comp.py
 pylj/tests/test_mc.py
 pylj/tests/test_md.py
 pylj/tests/test_pairwise.py
