### 2018-07-03 f46a1d8
adding try and except to test_comp
