+++ pylj/tests/test_comp.py
@@ -1,5 +1,10 @@
 from numpy.testing import assert_almost_equal, assert_equal
-from pylj import comp, util
+from pylj import util
+try:
+    from pylj import comp as heavy
+except ImportError:
+    print("WARNING, using slow force and energy calculations")
+    from pylj import pairwise as heavy
 import unittest
 import numpy as np
 
@@ -9,7 +14,7 @@ class TestPairwise(unittest.TestCase):
         particles = np.zeros(2, dtype=part_dt)
         particles['xposition'][0] = 1e-10
         particles['xposition'][1] = 5e-10
-        particles, distances, forces, energies = comp.compute_forces(particles, 30, 15)
+        particles, distances, forces, energies = heavy.compute_forces(particles, 30, 15)
         assert_almost_equal(distances, [4e-10])
         assert_almost_equal(energies, [-1.4515047e-21])
         assert_almost_equal(forces, [-9.5864009e-12])
@@ -22,7 +27,7 @@ class TestPairwise(unittest.TestCase):
         particles = np.zeros(2, dtype=part_dt)
         particles['xposition'][0] = 1e-10
         particles['xposition'][1] = 5e-10
-        distances, energies = comp.compute_energy(particles, 30, 15)
+        distances, energies = heavy.compute_energy(particles, 30, 15)
         assert_almost_equal(distances, [4e-10])
         assert_almost_equal(energies, [-1.4515047e-21])
 
@@ -31,5 +36,5 @@ class TestPairwise(unittest.TestCase):
         particles = np.zeros(2, dtype=part_dt)
         particles['xposition'][0] = 1e-10
         particles['xposition'][1] = 5e-10
-        pressure = comp.calculate_pressure(particles, 30, 300, 15)
+        pressure = heavy.calculate_pressure(particles, 30, 300, 15)
         assert_almost_equal(pressure*1e24, 7.07368869)
