+++ pylj/tests/test_util.py
@@ -13,7 +13,7 @@ class TestUtil(unittest.TestCase):
         assert_almost_equal(a.particles['yposition'] * 1e10, [2, 6])
         assert_almost_equal(a.initial_particles['xposition'] * 1e10, [2, 2])
         assert_almost_equal(a.initial_particles['yposition'] * 1e10, [2, 6])
-        assert_almost_equal(a.cut_off * 1e10, 15)
+        assert_almost_equal(a.cut_off * 1e10, 4.0)
         assert_equal(a.distances.size, 1)
         assert_equal(a.forces.size, 1)
         assert_equal(a.energies.size, 1)
@@ -28,7 +28,7 @@ class TestUtil(unittest.TestCase):
         self.assertTrue(0 <= a.particles['yposition'][0] * 1e10 <= 8)
         self.assertTrue(0 <= a.particles['xposition'][1] * 1e10 <= 8)
         self.assertTrue(0 <= a.particles['yposition'][1] * 1e10 <= 8)
-        assert_almost_equal(a.cut_off * 1e10, 15)
+        assert_almost_equal(a.cut_off * 1e10, 4.0)
         assert_equal(a.distances.size, 1)
         assert_equal(a.forces.size, 1)
         assert_equal(a.energies.size, 1)
