+++ docs/source/using-pylj.rst
@@ -4,7 +4,7 @@ using pylj
 There are a number of ways to get using pylj:
 
 - Fork the code: please feel free to fork the code on GitHub_ and add functionality that interests you. 
-- Run it locally: pylj is available through the pip package manager. However, it does require a Jupyter notebooks environment. Therefore we recommend using the anaconda_ package, which gives Jupyter notebooks as well as packages such as numpy and matplotlib. Using a Windows computer, it may also be necessary to install a C++ compiler, usually the Visual C++ package_ is best.
+- Run it locally: pylj is available through the pip package manager. However, it does require a Jupyter notebooks environment. Therefore we recommend using the anaconda_ package, which gives Jupyter notebooks as well as packages such as numpy and matplotlib. For best performance a C++ compiler is required, usually the Visual C++ package_ is best (Windows).
 - Use the web app: pylj is available online at the following link_.
 - Get in touch: Andrew_ is always keen to chat to potential users or educators so feel free to drop him an email. 
 
