+++ docs/source/pairwise.rst
@@ -1,0 +1,9 @@
+pylj\.comp
+==========
+
+Generally these are functions required by pylj which involve a pair-wise comparison of particles. This is the python implimentation of the :doc:`comp` module.
+
+.. automodule:: pylj.pairwise
+    :members:
+    :undoc-members:
+    :show-inheritance:
