+++ docs/source/pairwise.rst
@@ -1,4 +1,4 @@
-pylj\.comp
+pylj\.pairwise
 ==========
 
 Generally these are functions required by pylj which involve a pair-wise comparison of particles. This is the python implimentation of the :doc:`comp` module.
