### 2018-07-05 arm61/pylj#11 @arm61
I have added the guide notebook which acts as a short instruction manual for the installation of pylj and introduces some of the examples.

Note that I have not (and don't plan to) put the long form examples on the web resource as the purpose of this resource is really to give a quick example.

I will close this issue (but feel free to re-open if unfulfilled)
