### 2018-07-06 #19 @labarba
I have an observation that is optional for you to address, but I ask you to consider… You are using British spelling in the `mc.initialise()` function. This is likely to trip many if not most users! I see that _you_ are based in Bath, so it make sense for your local users. Maybe you can overload the name and have two functions that do the same thing?
