### 2018-07-06 2697140
addition of the initialize maps
