### 2018-07-06 271a53a
update paper to mention all examples
