### 2018-07-06 21baa24
adding references to gromacs, lammps and dlpoly
