+++ paper/paper.bib
@@ -76,3 +76,44 @@ month = "6",
 url = "https://figshare.com/articles/fig1/6571583",
 doi = "10.6084/m9.figshare.6571583.v1"
 }
+
+@article{GROMACS,
+title = "GROMACS: A message-passing parallel molecular dynamics implementation",
+journal = "Computer Physics Communications",
+volume = "91",
+number = "1",
+pages = "43 - 56",
+year = "1995",
+issn = "0010-4655",
+doi = "https://doi.org/10.1016/0010-4655(95)00042-E",
+url = "http://www.sciencedirect.com/science/article/pii/001046559500042E",
+author = "H.J.C. Berendsen and D. van der Spoel and R. van Drunen",
+keywords = "Molecular dynamics, Parallel computing"
+}
+
+@article{DLPOLY,
+author = { W.   Smith  and  C.W.   Yong  and  P.M.   Rodger },
+title = {DL\_POLY: Application to molecular simulation},
+journal = {Molecular Simulation},
+volume = {28},
+number = {5},
+pages = {385-471},
+year  = {2002},
+publisher = {Taylor & Francis},
+doi = {10.1080/08927020290018769},
+URL = {https://doi.org/10.1080/08927020290018769},
+eprint = {https://doi.org/10.1080/08927020290018769}
+}
+
+@article{LAMMPS,
+title = "Fast Parallel Algorithms for Short-Range Molecular Dynamics",
+journal = "Journal of Computational Physics",
+volume = "117",
+number = "1",
+pages = "1 - 19",
+year = "1995",
+issn = "0021-9991",
+doi = "https://doi.org/10.1006/jcph.1995.1039",
+url = "http://www.sciencedirect.com/science/article/pii/S002199918571039X",
+author = "Steve Plimpton"
+}
