+++ paper/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 # Summary
 
-pylj is an educational software designed to introduce students to classical simulation using a Lennard-Jones potential model.[@LENNARD-JONES] Currently, pylj is capable of performing both molecular dynamics (NVE and NVT ensembles) and Monte-Carlo (NVT ensemble) simulations, for a two dimensional argon gas system. Written in Python (with Cythonic pairwise functions), visualisation occurs through a Jupyter notebook,[@KLUYVER] making use of the matplotlib library (an example of the Interactions visualisation environment is shown below).[@HUNTER]  This allows for easy deployment within a computational laboratory setting and immediate student interaction without an explaination of the command line. Example notebooks that show how pylj can be used to simulate both molecular dynamics and Monte-Carlo are available from the GitHub repository, although a large variety of possible applications are possible.
+pylj is an educational software designed to introduce students to classical simulation using a Lennard-Jones potential model.[@LENNARD-JONES] Currently, pylj is capable of performing both molecular dynamics (NVE and NVT ensembles) and Monte-Carlo (NVT ensemble) simulations, for a two dimensional argon gas system. Written in Python (with Cythonic pairwise functions), visualisation occurs through a Jupyter notebook,[@KLUYVER] making use of the matplotlib library (an example of the Interactions visualisation environment is shown below).[@HUNTER]  This allows for easy deployment within a computational laboratory setting and immediate student interaction without an explaination of the command line, as is the case in other open-source molecular dynamics packages.[@GROMACS][@LAMMPS][@DLPOLY] Example notebooks that show how pylj can be used to simulate both molecular dynamics and Monte-Carlo are available from the GitHub repository, although a large variety of possible applications are possible.
 
 ![](fig1.png)
 *Figure 1. An example of the Interactions sampling class during a molecular dynamics simulation. [@FIGURE1]*
