### 2018-07-06 #19 @arm61
@TJFord I have added references to the paper to gromacs, lammps and dlpoly in https://github.com/arm61/pylj/commit/21baa244c887c5ced60d227ecd0bc30967ac104e
