+++ docs/source/pairwise.rst
@@ -1,5 +1,5 @@
 pylj\.pairwise
-==========
+==============
 
 Generally these are functions required by pylj which involve a pair-wise comparison of particles. This is the python implimentation of the :doc:`comp` module.
 
