### 2018-07-06 #19 @TJFord
@labarba I think the author @arm61 has addressed all my concerns. I suggest that it is ready for publication.
