### 2018-07-11 #19 @arm61
@labarba it seems that both reviewers are happy. I was wondering what the next stage was?
