### 2018-07-14 #19 @labarba
@arm61 Could you try moving the positioning of your figure down, so that more text moves to the first page, under Summary? There's a big chunk of white space there.
