### 2018-07-14 #19 @labarba
*Editorial feedback*

_Summary: _
(l.1) classical simulation—> classical atomistic simulation
(l.2) use bibles for the citation (and elsewhere)
(l.4) define NVE and NVT
(l.5) capitalize Argon
(l.7) capitalize Matplotlib
 (l.9) explaination—>explanation (typo)
This whole passage is a little hard to parse. The construction is perhaps a bit ritzy. Can you simplify?Something like:
PyLJ is written in Python (using Cython for pairwise interactions) and uses Jupyter notebooks and Matplotlib for visualization (example below). It can be easily deployed in a computer laboratory, and students interact with it without needing to use the command line, like they would when using other molecular dynamics software like Gromacs, DL_poly or others. (Refs.) We provide example notebooks in the repository, showing how to use PyLJ to simulate a 2D gas system using molecular dynamics and Monte-Carlo methods. A variety of other applications are possible.
(p.2) “thermostating” is jargon: please define or rephrase

_Statement of need:_
classical simulation—> classical atomistic simulation
PyLJ allows this by offering—> PyLJ offers
capitalize Matplotlib
“In addition to use in the introduction of the simulation itself” —> huh? Please rephrase.
—.
Audience: you mention in the first sentence undergraduates in chemistry and physics. Can you say more about where in the curriculum this would be used? (what course? what year?)
