### 2018-07-15 #19 @labarba
Sorry to give you comments piecemeal… I did the first part on a flight yesterday.

You give a short description of the examples, with a link to the folder in the repo, within the “Statement of Need” section. It doesn’t quite belong there. You could use a new heading “Usage” and talk about the examples there. It would also be good to add a bulleted list of the examples in the README. Navigating down the folders to find the examples is not user-friendly, and since some folders have two notebooks, a user may not know where to start. I, myself, click through the various notebooks, and don’t know what order one is supposed to go through this.

What I’d like to see is a paragraph in the paper (and maybe in a new README in the `examples` folder) that orients an independent learner or instructor on getting started and adopting this tool. It took me _many_ visits to your repository to finally stumble on the `guide.ipynb` file just now! Please add a link to this file from the README. I think the “How to teach with PyLJ” section here would be useful to have in the paper. (But please don’t use the un-semantic hyperlinked “here.”)
