### 2018-07-15 #19 @labarba
I went to the [Getting Started](http://pythoninchemistry.org/getting-started/) page on the documentation, and clicked on the Tutorial links, which launch a Binder session. However, they all give a "cannot resolve ref" error.
