### 2018-07-15 8d6ebd3
Update paper.md
