### 2018-07-15 #19 @labarba
Maybe the problem with the citation to [LENNARD-JONES](https://github.com/arm61/pylj/blob/8d6ebd3a02049f2290249648e4a988ac7d82ad0e/paper/paper.bib#L57) is that the entry has no author field.
