### 2018-07-15 94a5a49
Update paper.md
