### 2018-07-15 ddf306f
Update paper.bib
