+++ paper/paper.bib
@@ -56,6 +56,7 @@ author = "Daan Frenkel and Berend Smit"
 
 @article {LENNARD-JONES,
 	title = {On the determination of molecular fields. {\textemdash}II. From the equation of state of a gas},
+	author = {Jones, J. E.},
 	volume = {106},
 	number = {738},
 	pages = {463--477},
