### 2018-07-15 22b926f
Update paper.md
