+++ paper/paper.md
@@ -31,15 +31,22 @@ bibliography: paper.bib
 
 # Summary
 
-pylj is an educational software designed to introduce students to classical simulation using a Lennard-Jones potential model.[@LENNARD-JONES] Currently, pylj is capable of performing both molecular dynamics (NVE and NVT ensembles) and Monte-Carlo (NVT ensemble) simulations, for a two dimensional argon gas system. Written in Python (with Cythonic pairwise functions), visualisation occurs through a Jupyter notebook,[@KLUYVER] making use of the matplotlib library (an example of the Interactions visualisation environment is shown below).[@HUNTER]  This allows for easy deployment within a computational laboratory setting and immediate student interaction without an explaination of the command line, as is the case in other open-source molecular dynamics packages.[@GROMACS][@LAMMPS][@DLPOLY] Example notebooks that show how pylj can be used to simulate both molecular dynamics and Monte-Carlo are available from the GitHub repository, although a large variety of possible applications are possible.
-
-The `md` module currently allows a Velocity-Verlet [@FRENKEL200263] NVE simulation, which can be adapted to NVT by allowing the thermostating via velocity rescaling. The `mc` module allows a typical NVT Metropolis Monte-Carlo simulation[@FRENKEL200223], however there is scope for future development of μVT, Grand-Canonical simulations.
+pylj is an educational software to introduce students to classical atomistic simulation using a Lennnard-Jones potential model.[@LENNARD-JONES] pylj is written in Python (using Cython for pairwise interactions) and uses Jupyter notebooks [@KLUYVER] and Matplotlib for visualisation (see example below).[@HUNTER] It can be easily deployed in a computer laboratory, and students interact with it without needing to use the command line, as would be the case for other molecular dynamics packages like Gromacs, LAMMPS, or DLPOLY.[@GROMACS][@LAMMPS][@DLPOLY] We provide example notebooks in the repoository, showing how to use pylj to simulated a 2D gas system using eaither molecular dynamics or Monte-Carlo methods. A variety of other applications are possible. 
 
 The latest release of the software can be installed via pip, or from source at https://github.com/arm61/pylj. Full documentation can be found at https://pylj.readthedocs.io.
 
 # Statement of Need
 
-As computational science becomes more and more important in the training of undergraduate chemistry and physics students, there is a larger need for easy to use, open-source, sustainable tools that enable student interaction with classical simulation. pylj allows this by offering a simple library of functions from which students can build their own molecular dynamics or Monte-Carlo simulation and easy visualisation, leveraging matplotlib and the Jupyter notebook framework. In addition to use in the introduction of the simulation itself, pylj can also be used to discuss material properties showing how they may be probed using simulation. A few examples of how pylj could be used in a teaching laboratory can be found in the [examples](https://github.com/arm61/pylj/tree/master/examples) directory. These include an introduction to each of molecular dynamics and Monte Carlo simulations algorithms as well as an example of how pylj could be used to investigate the ideal gas law and the deviation from it at high particle densities. pylj enables easy student interaction with these commonplace simulation methods in a fast and open-source way, while offering educators tools to build custom visualisation environments allowing attention to be drawn to areas of specific focus for a given application.
[…]
