### 2018-07-15 #19 @arm61
Okay, I think that is everything that you've mentioned resolved. Hopefully, the paper is more suitable not (commits https://github.com/arm61/pylj/commit/8d6ebd3a02049f2290249648e4a988ac7d82ad0e https://github.com/arm61/pylj/commit/94a5a49d57cb76e94508bb40b2ee3300e5b03ed1 https://github.com/arm61/pylj/commit/ddf306f3896bc02d6b18555bfa01c1ee0844daa4 https://github.com/arm61/pylj/commit/22b926f1ecfc26b95269cdefd98fe3e427dafdb2). The only thing that you suggested that I haven't done is captalising argon, this is because (as I understand) it is IUPAC convention to not trade element names as proper nouns (https://chemistry.stackexchange.com/questions/6381/why-do-people-often-capitalize-element-names).

I have also added a mention of the examples in the README (commit https://github.com/arm61/pylj/commit/666951580418427cc863f7028a39547b5794e8be). I don't go into detail as my plan is to mention the paper once it is published which gives more details.

I have move the "How to teach with pylj" bit from the guide.ipynb file to the "Usage" section of the paper and as a result have decided that the guide.ipynb file is surplus to requirements.

Thanks for the thorough comments @labarba
