### 2018-07-15 d5f0a23
Update paper.md
