### 2018-07-15 #19 @labarba
I need to do a little local processing to publish your paper, but will have to do that tomorrow, due to commitments now.
