### 2018-07-15 #19 @labarba
OK. We'll wait. If the DOI does not resolve tomorrow morning your time (please check), then send an email to Zenodo help. I won't publish the JOSE paper until the DOI of the archive resolves.
