### 2018-07-15 #19 @labarba
I mean, at the moment it _is_ pushed to [JOSE-papers](https://github.com/openjournals/jose-papers), but I haven't made a Crossref deposit.
