### 2018-07-16 #19 @labarba
I checked and the DOI still gives me an error: https://doi.org/10.5281/zenodo.1312617
