### 2018-07-16 #19 @labarba
Crossref threw citation errors. Please fix DOI links for the Gromacs paper and Plimpton 1995.

I will have to do the whole process again!
