### 2018-07-16 ce91689
Update paper.bib
