+++ paper/paper.bib
@@ -86,7 +86,7 @@ number = "1",
 pages = "43 - 56",
 year = "1995",
 issn = "0010-4655",
-doi = "https://doi.org/10.1016/0010-4655(95)00042-E",
+doi = "10.1016/0010-4655(95)00042-E",
 url = "http://www.sciencedirect.com/science/article/pii/001046559500042E",
 author = "H.J.C. Berendsen and D. van der Spoel and R. van Drunen",
 keywords = "Molecular dynamics, Parallel computing"
@@ -114,7 +114,7 @@ number = "1",
 pages = "1 - 19",
 year = "1995",
 issn = "0021-9991",
-doi = "https://doi.org/10.1006/jcph.1995.1039",
+doi = "10.1006/jcph.1995.1039",
 url = "http://www.sciencedirect.com/science/article/pii/S002199918571039X",
 author = "Steve Plimpton"
 }
