### 2018-07-16 #19 @arm61
Resolved in https://github.com/arm61/pylj/commit/ce9168990facc09b3f2441e342902d135c2bae7e

Sorry about that, apparently the bibtex served by sciencedirect incorrectly prepends "http://doi.org/" for the doi...
