### 2018-07-16 #19 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
