### 2018-07-16 #19 @labarba
@arm61  -- your paper is now published in JOSE ‼️  Congratulations 🎉
You can now add the happy green JOSE button on your repository.

@shivupa, @TJFord — Thank you again for your review 🙏
